Checkpointing initial work on adding support for type expressions in function parameters. As I just want to get something working end-to-end first, I am making minimal changes to the grammar at this point in time. Specifically, for now, I am only allowing type expressions that are also identifiers in CoffeeScript, so that includes "string" and "number", but excludes "goog.dom.DomHelper" (trying to define a ParamVar as "Namespace Identifier" did not go over well with Jison). As part of this change, I had to modify how the rewriter adds implicit parentheses: it no longer tries to add them within a PARAM_START/PARAM_END pair. This was problematic because the type expression "number" in the following example:
sqrt = (number n) -> Math.sqrt(n)

Would normally be rewritten as the following by addImplicitParentheses in rewriter.coffee:

sqrt = (number(n)) -> Math.sqrt(n)

This happens because the rewriter interprets the series of identifiers as a function call (which produces a series of tokens that does not parse). Disabling this behavior in a param list appears to fix the problem.

Finally, I also had to reorganize how JSDoc is included in the generated JavaScript because there are too many places in the code where new lines of JSDoc could be introduced. Currently, it is only inserted in the correct place for a constructor, but not for a method. For example, from the following CoffeeScript:

class example.people.Person extends example.people.AbstractPerson

  constructor: (string first, string last) ->
    super()
    @first = first
    @last = last

The following JavaScript is generated, which is desired:

/**
 * @param {string} first
 * @param {string} last
 * @constructor
 * @extends {example.people.AbstractPerson}
 */
example.people.Person = function(first, last) {
  example.people.AbstractPerson.call(this);
  this.first = first;
  this.last = last;
};

However, when the "string" type expression is included on an instance method:

  setFirst: (string first) -> @first = first

the generated JavaScript is somewhat awkward:

example.people.Person.prototype.setFirst = /**
 * @param {string} first
 */
function(first) {
  return this.first = first;
};

Clearly, there is still plenty of work to do!